Only connect to the items window signals if it has a window
If the styleitem was created in a control without a window, it
failed to connect to the activeChanged() signal of QQuickWindow.
It would consequently warn about invalid null parameter to
QObject::connect()

If the control was added to a window at a later point, it would also
not connect to the activeChanged() signal of QQuickWindow. This could
cause that the control was rendered as it would be rendered in an
inactive window, even if the window actually was active.

Task-number: QTBUG-88553
Pick-to: 6.0
Change-Id: I3aa5948e150f0f4baa204943ec43ea8922421e75
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>