ScrollView: let flickable have pixelAligned set to true
A ScrollView will typically contain other controls. And
if we're using a dpr with a fraction (e.g 1.25), the
position of the controls will sometimes jitter as you
scroll. The result is that if e.g a Frame is wrapped
tight around around a Rectangle, the frame will sometimes
be outside the rectangle, and other times on top. This
is because of rounding issues when drawing the Frame
vs the Rectangle.

So let the default flickable used by a ScrollView have
pixelAligned set to true. This will give the best
cross-platform "out of the box" experience.
The developer can always choose to use his own Flickable
for ScrollView, and tweak it, if such default
behavior is not wanted.

Pick-to: 6.0
Change-Id: I7dc7b0b390dbf055f35cca2a6c15a6075e33c0a4
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>