Revert "Set the stencil buffer zone" and "Add clipNext null pointer guard"
This reverts commit 576d3fc2c84f324c035aa924c386f0615002fa16.
This reverts commit 6382ec9a6c961f2b7d7139659c11a047f6b68b07.

The original patch is broken in multiple ways:

- Any at least two level deep clip where rotation is involved results
in a failing assertion (triggers in debug builds only).

- The assertion looks legit: using calculations meant only for
90, 180, 270 degree rootations for arbitrary rotations seems wrong.

Taking frame captures with RenderDoc show that the generated scissor
rectangle that is set before the glClear for the stencil buffer is
_not_ correct, i.e. it does not correctly cover all the area that will
be touched in the stencil buffer in the following draw call.

- The logic for accessing the clip list is dubious. Hence a suspicion
that the patch is not correct when it comes to the handling of the
clip list: It is not clear why accessing clip->clipList()->matrix()
is correct when the clip list is more than 2 levels deep. At the same
time a single clipping item will not trigger the assertion because it
does not set the scissor before the clear at all (which is contrary to
what the patch is aiming to achieve).

The original patch has a follow-up crash fix that is incorrect in its
own right: it adds a null check for clipNext (i.e. the clip->clipList()
mentioned above), while including code in the true branch that should
execute regardless.

Fixes: QTBUG-89898
Task-number: QTBUG-83108
Change-Id: Ic28bd8670a7663fd6232aaf5739b05fd16d2da24
Reviewed-by: Dominik Holland <dominik.holland@qt.io>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>