Use QDateTime as internal name for the builtin date type
The internal name serves as hint to the C++ class. This is the same
as the internal name of url, for example. The internal name is not
exposed to QML. So, from a language perspective, we can choose any name
there.

All the JavaScript methods of the Date prototype are not actually
available on QDateTime. Due to the magic conversions the QML engine does
in the background, much of them work in practice. Yet, when generating
C++ code, we would need another conversion step to make them available.
In order to express part of this, we introduce a second type for the
JavaScript Date object. In the future we might link them via some
"converts to" property. Then tools could know that creating a
QJSManagedValue from a QDateTime (or QUrl etc) results in a richer API
with additional methods.

Change-Id: I13f66a37d985dbd3581d8bb84937adbbc31286fd
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Andrei Golubev <andrei.golubev@qt.io>