Drop JavaScript root object members from builtins.qmltypes
Those are not actually accessible from C++, and the description we had
was incomplete and misleading. There was no way to know that the "Math"
type, for example, was the type of a member of the global object called
"Math", but the "float" type was not. Furthermore the types given for
parameters and return values were only approximations. JavaScript's type
coercion allows us to use a number of other types, too.

There certainly is a place for a proper description of the global
object, but such a description should be generated from the actual
JavaScript root object as found in an actual QJSEngine. Furthermore, we
do need better syntax to describe it.

Change-Id: Ia573436df04ca967381e1e13dbd42a70bcc60979
Reviewed-by: Andrei Golubev <andrei.golubev@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>