QQmlIRLoader: Actually load RequiredPropertyExtraData
If a QML component wants to mark properties of its "parent" component as
required, it can do so via
required propertyName
The information about those properties is stored in a
RequiredPropertyExtraData data structure. This structure is already
serialized to disk in the QQmlIRWriter. However, we neglected to restore
it so far in the loader.

Fixes: QTBUG-90538
Change-Id: I789daff9bc881e4f35c942c77f5116b5284de81b
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit b322a971f06823a4356f2b3aa331501aa4d0dc7f)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>