Disentangle QQmlPropertyBinding
Having a vpointer in front of QQmlPropertyBinding is problematic
considering how often we convert to and from quintptr. This leads to
writing garbage into the vpointer when calling ref/unref, as those
assume the first thing at the start of the object is actually the
refcount.
Avoid those issues by using "composition" instead of inheritance.

Task-number: QTBUG-89505
Change-Id: Ida9cf83b53e60a8ea9b7f329568deeca3346148b
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>