Add "QML compiler" tests
Introduce some basic tests for C++ objects as if they are compiled with
QML compiler. Primary usage of this test suite is to make sure that
the interaction between C++ and QML engine works. The following cases
are tested:
- C++ bindings that go through the qml engine
- Signal handlers
- JS - QML - C++ bindings ecosystem
- Property aliases
- Property change handlers
- (Indirectly) visibility of C++ stuff in JavaScript/QML

Nested objects are ignored for now - it's a lot more code for fairly
little gain at this point, so it seems easier to ignore this until
there's a QML compiler that can generate them

Task-number: QTBUG-91039
Change-Id: I6949670c9de8a28c5cca814cf47f501f404800bd
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>