QQmlPropertyValueInterceptor: Add binding interception support
With the new property system, it is not enough to only intercept writes;
we also need the ability to intercept calls to the bindable interface.
This is done by adding a new bindable virtual method
QQmlPropertyValueInterceptor. The default implementation does not do any
interception at all for now. This allows us to keep the existing
interceptors for now, and to port them step by step. Once all
interceptors in Qt are ported, we can make the method in the interface
pure virtual.

Task-number: QTBUG-90999
Change-Id: I697658a1cd9a5204805a444e0d949213ba71b91c
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>