QHashedString(Ref): Use QStringView internally
We cannot store a QStringView internally, as that would increase the
size of our objects (QHashedString(Ref) does not support strings with
sizes that do not fit into 32bit, thus it can store the hash and the
size in one 64 bit block). What we can however do is to construct a
QStringView on demand (which is cheap), and use its methods.

The benefits are twofold: We get rid of some accidental complexity in
qtdeclarative; and the code in QStringView is actually SIMD optimized.

Change-Id: I2445a2d5a16b253f4971d7f3be0e1b274326eacb
Reviewed-by: Maximilian Goldstein <max.goldstein@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>